Remove the Servers projects and all code dependent on them
This includes moving the ServerConfiguration to the main Telemachus dll,
as it is potentially the only part we want to save (I certainly don't have
a hankering to rewrite all the configuration!).
