--- Servers/.gitignore
@@ -1,3 +1,0 @@
-bin
-obj
-*.suo
