--- Servers/Servers.sln
@@ -1,26 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Servers", "Servers\Servers.csproj", "{57E9F7C0-DB25-4B6D-AA53-C44D95B47468}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServersTest", "ServersTest\ServersTest.csproj", "{DEF5477E-CE26-45DF-8316-F762D7F417C0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{57E9F7C0-DB25-4B6D-AA53-C44D95B47468}.Debug|x86.ActiveCfg = Debug|x86
-		{57E9F7C0-DB25-4B6D-AA53-C44D95B47468}.Debug|x86.Build.0 = Debug|x86
-		{57E9F7C0-DB25-4B6D-AA53-C44D95B47468}.Release|x86.ActiveCfg = Release|x86
-		{57E9F7C0-DB25-4B6D-AA53-C44D95B47468}.Release|x86.Build.0 = Release|x86
-		{DEF5477E-CE26-45DF-8316-F762D7F417C0}.Debug|x86.ActiveCfg = Debug|x86
-		{DEF5477E-CE26-45DF-8316-F762D7F417C0}.Debug|x86.Build.0 = Debug|x86
-		{DEF5477E-CE26-45DF-8316-F762D7F417C0}.Release|x86.ActiveCfg = Release|x86
-		{DEF5477E-CE26-45DF-8316-F762D7F417C0}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
