--- Servers/Servers/Resources/ResponseCodePageTemplate.txt
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>$status $code</title>
-  </head>
-  <body>
-	<h1>$status $code</h1>
-    <pre>$reason</pre>
-  </body>
-</html>
