+++ Telemachus/src/TelemachusBehaviour.cs
@@ -1,6 +1,5 @@
 ﻿//Author: Richard Bunt
 using KSP.IO;
-using Servers.MinimalHTTPServer;
 using System;
 using System.Collections.Generic;
 using System.Reflection;
@@ -136,19 +135,6 @@ namespace Telemachus
             else
             {
                 PluginLogger.print("No IP address in configuration file.");
-            }
-
-
-            serverConfig.maxRequestLength = config.GetValue<int>("MAXREQUESTLENGTH");
-
-            if (serverConfig.maxRequestLength < 8000)
-            {
-                PluginLogger.print("No max request length specified, setting to 8000.");
-                serverConfig.maxRequestLength = 10000;
-            }
-            else
-            {
-                PluginLogger.print("Max request length set to:" + serverConfig.maxRequestLength);
             }
             
             serverConfig.version = System.Reflection.Assembly.GetExecutingAssembly().GetName().Version.ToString();
@@ -346,7 +332,7 @@ namespace Telemachus
         private PluginManager _manager;
 
         public KSPAPI(FormatterProvider formatters, VesselChangeDetector vesselChangeDetector,
-            Servers.AsynchronousServer.ServerConfiguration serverConfiguration, PluginManager manager)
+            ServerConfiguration serverConfiguration, PluginManager manager)
         {
             _manager = manager;
