+++ TelemachusTest/src/DummyKSPAPI.cs
@@ -9,7 +9,7 @@ namespace TelemachusTest
     public class DummyKSPAPI : IKSPAPI
     {
         public DummyKSPAPI(FormatterProvider formatters, VesselChangeDetector vesselChangeDetector,
-            Servers.AsynchronousServer.ServerConfiguration serverConfiguration)
+            ServerConfiguration serverConfiguration)
         {
             APIHandlers.Add(new DummyHandler(formatters));
         }
